+++ poem-openapi/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [2.0.7]
+# [2.0.7] 2022-08-02
 
 - Expose `AttachmentType` enum [#344](https://github.com/poem-web/poem/issues/344)
 - Add `rename_all` attribute for `Union` macro [#347](https://github.com/poem-web/poem/issues/347)